% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{latest}
\alias{latest}
\title{latest}
\usage{
latest(q, lim, rDrive)
}
\arguments{
\item{q}{A string.}

\item{lim}{An integer to limit number of tweets scrape.}

\item{rDrive}{An object of RSelenium driver.}
}
\value{
A tibble which is automatically stored in latest_tweet.rda
}
\description{
latest
}
