% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{people}
\alias{people}
\title{people}
\usage{
people(q, lim, rDrive)
}
\arguments{
\item{q}{A string.}

\item{lim}{An integer to limit number of tweets scrape.}

\item{rDrive}{An object of RSelenium driver.}
}
\value{
A tibble which is automatically stored in people_tweet.rda
}
\description{
people
}
