% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.R
\name{top}
\alias{top}
\title{top}
\usage{
top(q, lim, rDrive)
}
\arguments{
\item{q}{A string.}

\item{lim}{An integer to limit number of tweets scrape.}

\item{rDrive}{An object of RSelenium driver.}
}
\value{
A tibble which is automatically stored in top_tweet.rda
}
\description{
top
}
